requirements.txt
Django==2.1.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
Jinja2==2.10
MarkupSafe==1.1.0
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
pytz==2018.7
